more bug fixes -- change and fixed did the same thing, reviewed list should only include incorrect items, unload exercise list removed the history change listener
Former-commit-id: 1702be1a94eb4fac2341408185436247913d46b4 [formerly 9f1574aa91aed394bd607eb25a37341cfd82714a] [formerly 67590572591c32836a5dc279af4659688a6b0e42 [formerly ca958b6a27df9965f1716699699afe3dd73278cf]]
Former-commit-id: 7e94b9ff3848a5e52e5d8d2559cdfad9ad8dced8 [formerly 9c05cece4ef55ee8f4f9f4781d10344225664439]
Former-commit-id: e2583c587f8d04c805518076fc1cff38e930febb

+++ src/mitll/langtest/client/custom/ReviewEditItem.java
@@ -94,20 +94,24 @@ public class ReviewEditItem<T extends ExerciseShell> extends EditItem<T> {
     }
 
     protected void doAfterEditComplete(PagingContainer<T> pagingContainer, String buttonName) {
-      exerciseList.forgetExercise(exerciseList.byID(newUserExercise.getID()));
-      if (!ul.remove(newUserExercise)) {
-        System.err.println("\n\n\ndoAfterEditComplete : error - didn't remove");
-      }
-      service.removeReviewed(newUserExercise.getID(), new AsyncCallback<Void>() {
-        @Override
-        public void onFailure(Throwable caught) {
+      if (buttonName.equals(FIXED)) {
+        exerciseList.forgetExercise(exerciseList.byID(newUserExercise.getID()));
+        if (!ul.remove(newUserExercise)) {
+          System.err.println("\n\n\ndoAfterEditComplete : error - didn't remove");
         }
+        service.removeReviewed(newUserExercise.getID(), new AsyncCallback<Void>() {
+          @Override
+          public void onFailure(Throwable caught) {
+          }
 
-        @Override
-        public void onSuccess(Void result) {
-          predefinedContentList.reload();
-        }
-      });
+          @Override
+          public void onSuccess(Void result) {
+            predefinedContentList.reload();
+          }
+        });
+      } else {
+        super.doAfterEditComplete(pagingContainer, buttonName);
+      }
     }
   }
 }
